+++ scripts/systems/pf1.js
@@ -1,8 +1,12 @@
-export class StealthyBaseEngine {
+import { Stealthy, StealthyBaseEngine } from '../stealthy.js';
+
+export class StealthyPF1 extends StealthyBaseEngine {
 
   constructor() {
     // Hook the relevant skills to capture spot and hidden test
     // results into effects on the actor.
+    super();
+    console.warn(`Stealthy for '${game.system.id}' is stubbed out, needs development`);
   }
 
   testStealth(visionSource, config) {
@@ -63,83 +67,4 @@ export class StealthyBaseEngine {
     effect.flags.stealthy = flag;
     await actor.updateEmbeddedDocuments('ActiveEffect', [effect]);
   }
-}
-
-export class Stealthy {
-
-  static CONSOLE_COLORS = ['background: #222; color: #80ffff', 'color: #fff'];
-  static socket;
-  static enableSpot = true;
-  static engine;
-
-  static log(format, ...args) {
-    const level = game.settings.get('stealthy', 'logLevel');
-    if (level !== 'none') {
-
-      function colorizeOutput(format, ...args) {
-        return [
-          `%cstealthy %c|`,
-          ...Stealthy.CONSOLE_COLORS,
-          format,
-          ...args,
-        ];
-      }
-
-      if (level === 'debug')
-        console.debug(...colorizeOutput(format, ...args));
-      else if (level === 'log')
-        console.log(...colorizeOutput(format, ...args));
-    }
-  }
-
-  //########### Socket Functions ##############
-
-  static async toggleSpotting(toggled) {
-    Stealthy.enableSpot = toggled;
-
-    if (!toggled && game.user.isGM) {
-      const label = game.i18n.localize('stealthy-spot-label');
-      for (let token of canvas.tokens.placeables) {
-        const actor = token.actor;
-        const spot = actor.effects.find(e => e.label === label);
-        if (spot) {
-          actor.deleteEmbeddedDocuments('ActiveEffect', [spot.id]);
-        }
-      }
-    }
-  }
-
-  static async getSpotting() {
-    return Stealthy.enableSpot;
-  }
-
-  static async UpdateOrCreateEffect({ label, actor, flag, makeEffect }) {
-    let effect = actor.effects.find(e => e.label === label);
-
-    if (!effect) {
-      // See if we can source from outside
-      const source = game.settings.get('stealthy', 'hiddenSource');
-      if (source === 'ce') {
-        await game.dfreds.effectInterface.addEffect({ effectName: label, uuid: actor.uuid });
-        effect = actor.effects.find(e => e.label === label);
-      }
-      else if (source === 'cub') {
-        await game.cub.applyCondition(label, actor);
-        effect = actor.effects.find(e => e.label === label);
-      }
-
-      // If we haven't found an ouside source, create the default one
-      if (!effect) {
-        effect = makeEffect(flag, source);
-        await actor.createEmbeddedDocuments('ActiveEffect', [effect]);
-        return;
-      }
-    }
-
-    effect = duplicate(effect);
-    effect.flags.stealthy = flag;
-    effect.disabled = false;
-    await actor.updateEmbeddedDocuments('ActiveEffect', [effect]);
-  }
-
 }